+++ api/index.js
@@ -94,9 +94,9 @@ async function server(args, config, cb) {
         throw new Error(err);
     }
 
-    const auth = new (require('./lib/auth').Auth)(pool);
+    const user = new (require('./lib/user'))(pool);
     const email = new (require('./lib/email'))();
-    const authtoken = new (require('./lib/auth').AuthToken)(pool);
+    const token = new (require('./lib/token'))(pool);
 
     const app = express();
     const router = express.Router();
@@ -203,7 +203,7 @@ async function server(args, config, cb) {
             }
 
             try {
-                req.auth = await authtoken.validate(authorization[1]);
+                req.auth = await token.validate(authorization[1]);
                 req.auth.type = 'token';
             } catch (err) {
                 return Err.respond(err, res);
@@ -266,7 +266,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /upload'),
         async (req, res) => {
             try {
-                await auth.is_flag(req, 'upload');
+                await user.is_flag(req, 'upload');
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -308,9 +308,9 @@ async function server(args, config, cb) {
         ...schemas.get('GET /user'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
-                res.json(await auth.list(req.query));
+                res.json(await user.list(req.query));
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -334,9 +334,9 @@ async function server(args, config, cb) {
         ...schemas.get('POST /user'),
         async (req, res) => {
             try {
-                const user = await auth.register(req.body);
+                const user = await user.register(req.body);
 
-                const forgot = await auth.forgot(user.username, 'verify');
+                const forgot = await user.forgot(user.username, 'verify');
 
                 if (args.email) await email.verify({
                     username: user.username,
@@ -372,9 +372,9 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'id');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
-                res.json(await auth.patch(req.params.id, req.body));
+                res.json(await user.patch(req.params.id, req.body));
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -397,7 +397,7 @@ async function server(args, config, cb) {
         ...schemas.get('GET /login/verify'),
         async (req, res) => {
             try {
-                res.json(await auth.verify(req.query.token));
+                res.json(await user.verify(req.query.token));
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -524,7 +524,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /login/forgot'),
         async (req, res) => {
             try {
-                const reset = await auth.forgot(req.body.user); // Username or email
+                const reset = await user.forgot(req.body.user); // Username or email
 
                 if (args.email) await email.forgot(reset);
 
@@ -554,7 +554,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /login/reset'),
         async (req, res) => {
             try {
-                return res.json(await auth.reset({
+                return res.json(await user.reset({
                     token: req.body.token,
                     password: req.body.password
                 }));
@@ -581,9 +581,9 @@ async function server(args, config, cb) {
         ...schemas.get('GET /token'),
         async (req, res) => {
             try {
-                await auth.is_auth(req);
+                await user.is_auth(req);
 
-                return res.json(await authtoken.list(req.auth));
+                return res.json(await token.list(req.auth));
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -607,9 +607,9 @@ async function server(args, config, cb) {
         ...schemas.get('POST /token'),
         async (req, res) => {
             try {
-                await auth.is_auth(req);
+                await user.is_auth(req);
 
-                return res.json(await authtoken.generate(req.auth, req.body.name));
+                return res.json(await token.generate(req.auth, req.body.name));
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -634,9 +634,9 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'id');
 
-                await auth.is_auth(req);
+                await user.is_auth(req);
 
-                return res.json(await authtoken.delete(req.auth, req.params.id));
+                return res.json(await token.delete(req.auth, req.params.id));
             } catch (err) {
                 return Err.respond(err, res);
             }
@@ -660,7 +660,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /schedule'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 await Schedule.event(pool, req.body);
 
@@ -726,7 +726,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'collection');
 
-                await auth.is_auth(req);
+                await user.is_auth(req);
 
                 Collection.data(pool, req.params.collection, res);
             } catch (err) {
@@ -755,7 +755,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'collection');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 await Collection.delete(pool, req.params.collection);
 
@@ -786,7 +786,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /collections'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const collection = new Collection(req.body.name, req.body.sources);
                 await collection.generate(pool);
@@ -820,7 +820,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'collection');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const collection = await Collection.from(pool, req.params.collection);
 
@@ -1012,7 +1012,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /run'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const run = await Run.generate(pool, req.body);
 
@@ -1096,7 +1096,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'run');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const run = await Run.from(pool, req.params.run);
 
@@ -1136,7 +1136,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'run');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const jobs = await Run.populate(pool, req.params.run, req.body.jobs);
 
@@ -1298,7 +1298,7 @@ async function server(args, config, cb) {
         ...schemas.get('POST /job/error'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const joberror = new JobError(req.body.job, req.body.message);
                 return res.json(await joberror.generate(pool));
@@ -1329,7 +1329,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'job');
 
-                await auth.is_flag(req, 'moderator');
+                await user.is_flag(req, 'moderator');
 
                 res.json(JobError.moderate(pool, ci, req.params.job, req.body));
             } catch (err) {
@@ -1385,7 +1385,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'job');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const job = await Job.from(pool, req.params.job);
                 const run = await Run.from(pool, job.run);
@@ -1483,7 +1483,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'job');
 
-                await auth.is_auth(req);
+                await user.is_auth(req);
 
                 await Job.data(pool, req.params.job, res);
             } catch (err) {
@@ -1543,7 +1543,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'job');
 
-                await auth.is_auth(req);
+                await user.is_auth(req);
 
                 Job.cache(req.params.job, res);
             } catch (err) {
@@ -1604,7 +1604,7 @@ async function server(args, config, cb) {
             try {
                 await Param.int(req, 'job');
 
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 const job = await Job.from(pool, req.params.job);
 
@@ -1637,7 +1637,7 @@ async function server(args, config, cb) {
         ...schemas.get('GET /dash/traffic'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 res.json(await analytics.traffic());
             } catch (err) {
@@ -1662,7 +1662,7 @@ async function server(args, config, cb) {
         ...schemas.get('GET /dash/collections'),
         async (req, res) => {
             try {
-                await auth.is_admin(req);
+                await user.is_admin(req);
 
                 res.json(await analytics.collections());
             } catch (err) {
