+++ cloudformation/api.js
@@ -7,6 +7,10 @@ const stack = {
             Type: 'String',
             Description: 'Secret for auth against internal API functions'
         },
+        OpenCollective: {
+            Type: 'String',
+            Description: 'OpenCollective API Token'
+        },
         MailGun: {
             Type: 'String',
             Description: 'MailGun API Token to send user emails'
@@ -209,6 +213,7 @@ const stack = {
                         { Name: 'ECS_LOG_LEVEL', Value: 'debug' },
                         { Name: 'MAPBOX_TOKEN', Value: cf.ref('MapboxToken') },
                         { Name: 'MAILGUN_API_KEY', Value: cf.ref('MailGun') },
+                        { Name: 'OPENCOLLECTIVE_API_KEY', Value: cf.ref('OpenCollective') },
                         { Name: 'POSTGRES', Value: cf.join(['postgresql://openaddresses:', cf.ref('DatabasePassword'), '@', cf.getAtt('DBInstance', 'Endpoint.Address'), ':5432/openaddresses']) },
                         { Name: 'SharedSecret', Value: cf.ref('SharedSecret') },
                         { Name: 'GithubSecret', Value: cf.ref('GithubSecret') },
